Add session and Neo4j

+++ wardrobe/api/delete-outfit.php
@@ -1,12 +1,19 @@
 <?php
-$id = $_GET['id'];
-// Connect
-include "../db-connect.php";
+session_start(); // Session
+$id = $_GET['id']; // Get id outfit
+include "../db-connect.php"; // Connect
 
-// Query update clothes
-$query = "DELETE FROM outfits WHERE id = '$id'";
-$result = mysql_query($query,$db);
-header('Location: ../outfits.php?');
+// MySQL
+if($_SESSION['db_mode']=="MySQL") {
+	// Query DELETE outfit
+	$query = "DELETE FROM outfits WHERE id = '$id'";
+	$result = mysql_query($query,$db);
+
+// Neo4j
+} else if($_SESSION['db_mode']=="Neo4j") {
+
+}
+
+// Back
+header('Location: ../outfits.php');
 ?>
-
-
